// src/components/button/src/types/types.ts
export type ButtonColors = 'primary' | 'secondary' | 'outline' | 'outline-black' | undefined;
export type ButtonSizes = 'lg' | 'sm';
export type ButtonFontWeight = 'regular' | 'medium' | 'bold' | undefined;
export type ButtonType = 'submit' | 'reset' | 'button' | undefined;

export enum Colors {
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
  OUTLINE = 'outline',
  OUTLINE_BLACK = 'outline-black',
}

export enum Weight {
  REGULAR = 'regular',
  MEDIUM = 'medium',
  BOLD = 'bold',
}